refactor: move config utils to its own folder
