+++ config/utils/copySync.js
@@ -1,0 +1,11 @@
+const fs = require('fs');
+
+const copySync = (src, dest, overwrite) => {
+  if (overwrite && fs.existsSync(dest)) {
+    fs.unlinkSync(dest);
+  }
+  const data = fs.readFileSync(src);
+  fs.writeFileSync(dest, data);
+};
+
+module.exports = copySync;
