+++ config/utils/createIfDoesntExist.js
@@ -1,0 +1,9 @@
+const fs = require('fs');
+
+const createIfDoesntExist = (dest) => {
+  if (!fs.existsSync(dest)) {
+    fs.mkdirSync(dest);
+  }
+};
+
+module.exports = createIfDoesntExist;
