+++ config/utils/index.js
@@ -1,0 +1,9 @@
+const copySync = require('./copySync');
+const copySyncIfDoesntExist = require('./copySyncIfDoesntExist');
+const createIfDoesntExist = require('./createIfDoesntExist');
+
+module.exports = {
+  copySync,
+  copySyncIfDoesntExist,
+  createIfDoesntExist
+};
